requirement.txt
django
djangorestframework
pandas
numpy